MNT: move utilities around
